requirements.txt

numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
matplotlib==3.8.0
pandas==2.1.1